도감_목록_상세_조회_API_구현 :   refactor : 도감 요약 API 응답(totalCount/collectedCount) 타입을 long으로 통일, 기존 CatalogItemListResponse와 일치시킴 https://github.com/eodaego/eodaego-BE/issues/12

+++ src/main/java/com/chuseok22/eodaegoserver/domain/catalog/dto/response/CatalogCategorySummaryResponse.java
@@ -9,10 +9,10 @@ public record CatalogCategorySummaryResponse(
     CatalogCategory category,
 
     @Schema(description = "이 카테고리의 전체 도감 항목 개수", example = "48")
-    int totalCount,
+    long totalCount,
 
     @Schema(description = "이 카테고리에서 현재 회원이 수집한 개수", example = "12")
-    int collectedCount,
+    long collectedCount,
 
     @Schema(description = "이 카테고리의 수집률(정수 퍼센트, 0~100)", example = "25")
     int collectionRate
